add more e2e cases
